Move `NotAuthorized` component into more general file
