+++ frontend/src/routes/manage/Realm/AddChild.tsx
@@ -14,8 +14,8 @@ import { useForm } from "react-hook-form";
 import { Input } from "../../../ui/Input";
 import { Form } from "../../../ui/Form";
 import { PathSegmentInput } from "../../../ui/PathSegmentInput";
-import { NotAuthorized, PathInvalid } from ".";
-import { boxError } from "../../../ui/error";
+import { PathInvalid } from ".";
+import { boxError, NotAuthorized } from "../../../ui/error";
 import { displayCommitError, RealmSettingsContainer, realmValidations } from "./util";
 import { Button } from "../../../ui/Button";
 import { AddChildMutationResponse } from "../../../query-types/AddChildMutation.graphql";
