+++ frontend/src/ui/error.tsx
@@ -1,4 +1,6 @@
 import { ReactNode } from "react";
+import { useTranslation } from "react-i18next";
+
 import { Card } from "./Card";
 
 
@@ -15,3 +17,8 @@ export const ErrorBox: React.FC = ({ children }) => (
 export const boxError = (err: ReactNode): JSX.Element | null => (
     err == null ? null : <ErrorBox>{err}</ErrorBox>
 );
+
+export const NotAuthorized: React.FC = () => {
+    const { t } = useTranslation();
+    return <ErrorBox>{t("errors.not-authorized-to-view-page")}</ErrorBox>;
+};
